Moved PaddleApplicationSettings to :core and removed duality
