+++ core/src/main/kotlin/io/paddle/plugin/standard/extensions/Locations.kt
@@ -1,12 +1,9 @@
 package io.paddle.plugin.standard.extensions
 
 import io.paddle.project.PaddleProject
-import io.paddle.utils.exists
+import io.paddle.utils.config.PaddleApplicationSettings
 import io.paddle.utils.ext.Extendable
 import java.nio.file.Path
-import java.nio.file.Paths
-import kotlin.io.path.createDirectories
-import kotlin.io.path.createFile
 
 val PaddleProject.locations: Locations
     get() = extensions.getOrFail(Locations.Extension.key)
@@ -19,14 +16,8 @@ class Locations private constructor(val project: PaddleProject) {
     }
 
     val paddleHome: Path
-        get() {
-            val fromEnv = project.executor.env.get("PADDLE_HOME")?.let { Path.of(it) }
-            val result = fromEnv ?: Paths.get(project.executor.os.userHome, ".paddle")
-            if (!result.exists()) result.createDirectories()
-            return result
-        }
+        get() = PaddleApplicationSettings.paddleHome
 
     val registry: Path
-        get() = paddleHome.resolve("registry.yaml")
-            .apply { if (!exists()) createFile() }
+        get() = PaddleApplicationSettings.registry
 }
