+++ idea/src/main/kotlin/io/paddle/idea/settings/global/PaddleAppSettingsConfigurable.kt
@@ -4,6 +4,7 @@ import com.intellij.openapi.options.BoundSearchableConfigurable
 import com.intellij.ui.dsl.builder.*
 import io.paddle.idea.settings.global.PaddleAppSettings.TaskTypeOnProjectReload.INSTALL
 import io.paddle.idea.settings.global.PaddleAppSettings.TaskTypeOnProjectReload.RESOLVE
+import io.paddle.utils.config.PaddleApplicationSettings
 
 class PaddleAppSettingsConfigurable : BoundSearchableConfigurable(
     displayName = "Paddle Application Settings",
@@ -26,10 +27,10 @@ class PaddleAppSettingsConfigurable : BoundSearchableConfigurable(
                 .bindSelected(PaddleAppSettings.getInstance()::isDontShowDialogOnRequirementTxtPaste)
         }
         row {
-            checkBox("Disable pip's caching behaviour (pass --no-cache-dir flag)").bindSelected(PaddleRegistryInteract.Python::noCacheDir)
+            checkBox("Disable pip's caching behaviour (pass --no-cache-dir flag)").bindSelected(PaddleApplicationSettings.Python::noCacheDir)
         }
         row {
-            checkBox("Replace cached local wheels with verified wheels of the same version from PyPI").bindSelected(PaddleRegistryInteract.Python::autoRemove)
+            checkBox("Replace cached local wheels with verified wheels of the same version from PyPI").bindSelected(PaddleApplicationSettings.Python::autoRemove)
         }
     }
 }
